api/requirements.txt
alembic==1.4.2
aniso8601==7.0.0
apache-libcloud==2.8.1
appdirs==1.4.3
arpy==1.1.1
attrs==19.3.0
audioread==2.1.8
autopep8==1.5
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
boto3==1.12.30
botocore==1.15.30
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
depot==0.0.12
docopt==0.6.2
docutils==0.15.2
filedepot==0.7.1
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Bcrypt==0.7.1
Flask-CLI==0.4.0
Flask-Cors==3.0.8
Flask-GraphQL==2.0.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
graphene==2.1.8
graphene-file-upload==1.2.2
graphene-sqlalchemy==2.2.2
graphene-sqlalchemy-filter==1.10.2
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jwt==0.6.1
librosa==0.7.2
llvmlite==0.31.0
lockfile==0.12.2
lxml==4.5.0
Mako==1.1.2
MarkupSafe==1.1.1
numba==0.48.0
numpy==1.18.2
pathspec==0.7.0
pendulum==2.1.0
promise==2.3
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-gnupg==0.4.5
pytz==2019.3
pytzdata==2019.3
PyYAML==5.3.1
regex==2020.2.20
requests==2.23.0
resampy==0.2.2
Rx==1.6.1
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
singledispatch==3.4.0.3
six==1.14.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.15
toml==0.10.0
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
Werkzeug==1.0.0
